+++ app/src/main/java/com/tellingus/tellingme/data/model/oauth/login/OauthRequestDto.kt
@@ -1,4 +1,4 @@
-package com.tellingus.tellingme.data.model.login
+package com.tellingus.tellingme.data.model.oauth.login
 
 data class OauthRequestDto(
     val socialId: String = "-1"
