+++ app/src/main/java/com/tellingus/tellingme/data/network/NetworkService.kt
@@ -1,7 +1,7 @@
 package com.tellingus.tellingme.data.network
 
-import com.tellingus.tellingme.data.model.oauth.dto.OauthRequestDto
-import com.tellingus.tellingme.data.model.oauth.response.TokenDto
+import com.tellingus.tellingme.data.model.oauth.login.OauthRequestDto
+import com.tellingus.tellingme.data.model.oauth.login.TokenDto
 import com.tellingus.tellingme.data.network.adapter.ApiResult
 import retrofit2.http.Body
 import retrofit2.http.Header
